+++ chainer/functions/math/erfcx.py
@@ -36,7 +36,7 @@ class Erfcx(function_node.FunctionNode):
     def forward_gpu(self, x):
         self.retain_inputs((0,))
         self.retain_outputs((0,))
-        return cuda.cupyx.scipy.special.erfcx(x[0]),
+        return cuda.cupyx.scipy.special.erfcx(x[0]).astype(x[0].dtype),
 
     def backward(self, indexes, gy):
         x = self.get_retained_inputs()[0]
